Conflicts:
	src/app/guards/auth.guard.ts

+++ src/app/services/room.service.ts
@@ -1,10 +1,13 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from "@angular/common/http";
+import { ParseArgumentException } from "@angular/cli/models/parser";
 import { Observable } from "rxjs";
 import { environment } from "../../environments/environment";
 
 import { Room } from "../types/room";
 import { RoomId } from "../types/aliases/room-id";
+import { RoomTimespan } from "../types/room-timespan";
+import { TimespanId } from "../types/aliases/timespan-id";
 
 @Injectable({
   providedIn: 'root'
@@ -59,6 +62,10 @@ export class RoomService {
    * @param {object} changedData changed fields of room
    */
   public editRoomData(roomId: RoomId, changedData: object): Observable<any> {
+    if(roomId === null) {
+      throw ParseArgumentException;
+    }
+
     const apiURL = `${environment.baseUrl}${environment.apiRoutes.editRoom.replace(':id', roomId)}`;
     const requestBody = {
       roomId: roomId,
@@ -74,10 +81,70 @@ export class RoomService {
    * @param {RoomId} roomId id of room
    */
   public deleteRoom(roomId: RoomId): Observable<any> {
+    if(roomId === null) {
+      throw ParseArgumentException;
+    }
+
     const apiURL = `${environment.baseUrl}${environment.apiRoutes.deleteRoom.replace(':id', roomId)}`;
 
     return this.httpClient.delete(apiURL);
   }
 
-  //@todo (un-)available times setRoomData
+  /**
+   * Creates timeslot where room is available, room is now bookable in this timeslot
+   *
+   * @param timeslot time
+   */
+  public createAvailableTimeslot(timeslot: RoomTimespan) {
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.createAvailableTimeslot}`;
+    const requestBody = {
+      timeslot: timeslot,
+    };
+
+    return this.httpClient.post(apiURL, requestBody);
+  }
+
+  /**
+   * Creates timeslot where room is unavailable, room is now not bookable in the timeslot
+   *
+   * @param timeslot time
+   */
+  public createUnavailableTimeslot(timeslot: RoomTimespan) {
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.createUnavailableTimeslot}`;
+    const requestBody = {
+      timeslot: timeslot,
+    };
+
+    return this.httpClient.post(apiURL, requestBody);
+  }
+
+  /**
+   * Deletes an available timeslot
+   *
+   * @param TimespanId id of timeslot
+   */
+  public deleteAvailableTimeslot(TimespanId: TimespanId): Observable<any> {
+    if(TimespanId === null) {
+      throw ParseArgumentException;
+    }
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.deleteAvailableTimeslot
+      .replace(':timeslot_id', TimespanId)}`;
+
+    return this.httpClient.delete(apiURL);
+  }
+
+  /**
+   * Deletes an unavailable timeslot
+   *
+   * @param TimespanId id of timeslot
+   */
+  public deleteUnavailableTimeslot(TimespanId: TimespanId): Observable<any> {
+    if(TimespanId === null) {
+      throw ParseArgumentException;
+    }
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.deleteUnavailableTimeslot
+      .replace(':timeslot_id', TimespanId)}`;
+
+    return this.httpClient.delete(apiURL);
+  }
 }
